@@ -14,31 +14,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.4.0",
-    "@angular/cdk": "~13.3.9",
-    "@angular/common": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/localize": "~13.4.0",
-    "@angular/material": "~13.3.9",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
-    "@ng-bootstrap/ng-bootstrap": "^11.0.0",
-    "@ng-select/ng-select": "^8.0.0",
-    "ng-zorro-antd": "13.4.0",
-    "ngx-cookie-service": "^13.2.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/localize": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@ng-bootstrap/ng-bootstrap": "^13.1.1",
+    "@ng-select/ng-select": "^9.1.0",
+    "ng-zorro-antd": "14.3.0",
+    "ngx-cookie-service": "^14.0.1",
     "@tinymce/tinymce-angular": "^5.0.0",
     "@types/geojson": "^7946.0.10",
     "@types/leaflet": "^1.7.9",
     "bootstrap": "^4.6.2",
     "jquery": "^3.6.4",
-    "popper.js": "^1.16.1",
+    "@popperjs/core": "^2.10.2",
     "classlist.js": "1.1.20150312",
     "core-js": "^3.22.8",
     "hammerjs": "^2.0.8",
     "jszip": "^3.10.1",
-    "keycloak-angular": "^9.3.0",
+    "keycloak-angular": "^12.2.0",
     "keycloak-js": "^25.0.6",
     "leaflet": "^1.7.1",
     "lodash": "^4.17.21",
@@ -55,17 +55,17 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-devkit/core": "~13.3.11",
-    "@angular-devkit/schematics": "13.3.11",
-    "@angular-eslint/builder": "^13.5.0",
-    "@angular-eslint/eslint-plugin": "^13.5.0",
-    "@angular-eslint/eslint-plugin-template": "^13.5.0",
-    "@angular-eslint/schematics": "^13.5.0",
-    "@angular-eslint/template-parser": "^13.5.0",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler": "~13.4.0",
-    "@angular/compiler-cli": "~13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-devkit/core": "^14.2.13",
+    "@angular-devkit/schematics": "14.2.13",
+    "@angular-eslint/builder": "^14.4.0",
+    "@angular-eslint/eslint-plugin": "^14.4.0",
+    "@angular-eslint/eslint-plugin-template": "^14.4.0",
+    "@angular-eslint/schematics": "^14.4.0",
+    "@angular-eslint/template-parser": "^14.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
     "@types/arcgis-js-api": "4.6.0",
     "@types/jasmine": "~3.10.0",
     "@types/jest": "^27.0.3",
