@@ -123,203 +123,129 @@ export function initConfig(configService: ConfigService, keycloakService: Keyclo
 }
 
 @NgModule({
-  declarations: [
-    ActivityComponent,
-    ActivityDetailTableRowsComponent,
-    ActivityTableRowsComponent,
-    AddCommentComponent,
-    AddDocumentComponent,
-    AddDocumentTableRowsComponent,
-    AddEditActivityComponent,
-    AddEditCommentPeriodComponent,
-    AddEditContactComponent,
-    AddEditOrganizationComponent,
-    AddEditProjectNotificationComponent,
-    AddLabelComponent,
-    AppComponent,
-    ApplicationSortTableRowsComponent,
-    CommentPeriodBannerComponent,
-    CommentPeriodComponent,
-    CommentPeriodDetailsTabComponent,
-    CommentPeriodsComponent,
-    CommentPeriodsTableRowsComponent,
-    ConfirmComponent,
-    ContactsComponent,
-    DayCalculatorModalComponent,
-    DocumentApplicationSortComponent,
-    DocumentDetailComponent,
-    DocumentEditComponent,
-    DocumentTableRowsComponent,
-    FooterComponent,
-    GroupContactComponent,
-    GroupContactSelectComponent,
-    GroupTableRowsComponent,
-    GroupsTableRowsComponent,
-    HeaderComponent,
-    HomeComponent,
-    InputModalComponent,
-    LoginComponent,
-    MapComponent,
-    MetricsComponent,
-    MilestonesComponent,
-    NotAuthorizedComponent,
-    OrganizationsComponent,
-    OrganizationsTableRowsComponent,
-    PinsListComponent,
-    PinsTableRowsComponent,
-    ProjectArchivedDetailComponent,
-    ProjectComponent,
-    ProjectDetailComponent,
-    ProjectDocumentsComponent,
-    ProjectGroupsComponent,
-    ProjectNotificationTableRowsComponent,
-    ProjectNotificationsComponent,
-    ProjectUpdatesComponent,
-    ReviewCommentComponent,
-    ReviewCommentsTabComponent,
-    ReviewCommentsTabTableRowsComponent,
-    SearchComponent,
-    SearchDocumentTableRowsComponent,
-    SearchHelpComponent,
-    SidebarComponent,
-    ToggleButtonComponent,
-    ProjectDocumentsUploadComponent,
-    UserTableRowsComponent,
-    ProjectNotificationUploadComponent,
-    ProjectNotificationDetailComponent,
-    ProjectNotificationComponent,
-    ProjectNotificationDocumentsComponent,
-    PnDocumentTableRowsComponent,
-    AddEditProjectComponent,
-    ProjectListComponent,
-    ProjectListTableRowsComponent,
-    FormTab2002Component,
-    FormTab2018Component,
-    ModificationsListTableRowsComponent
-  ],
-  imports: [
-    AppRoutingModule, // <-- module import order matters - https://angular.io/guide/router#module-import-order-matters
-    BrowserAnimationsModule,
-    BrowserModule,
-    CommonModule,
-    EditorModule,
-    FormsModule,
-    HttpClientModule,
-    MatMenuModule,
-    MatSlideToggleModule,
-    MatTabsModule,
-    NgbModule,
-    NgSelectModule,
-    NzPaginationModule,
-    ReactiveFormsModule,
-    SharedModule,
-  ],
-  providers: [
-    ApiService,
-    CanDeactivateGuard,
-    CommentPeriodService,
-    CommentService,
-    ConfigService,
-    CookieService,
-    CurrencyPipe,
-    DecisionService,
-    DocumentService,
-    ExcelService,
-    KeycloakService,
-    NotificationProjectService,
-    ProjectService,
-    RecentActivityService,
-    SearchService,
-    SideBarService,
-    StorageService,
-    UserService,
-    {
-      provide: APP_INITIALIZER,
-      useFactory: initConfig,
-      deps: [ConfigService, KeycloakService],
-      multi: true
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: TokenInterceptor,
-      multi: true
-    },
-  ],
-  entryComponents: [
-    ActivityComponent,
-    ActivityDetailTableRowsComponent,
-    ActivityTableRowsComponent,
-    AddCommentComponent,
-    AddDocumentComponent,
-    AddDocumentTableRowsComponent,
-    AddEditActivityComponent,
-    AddEditCommentPeriodComponent,
-    AddEditContactComponent,
-    AddEditOrganizationComponent,
-    AddEditProjectNotificationComponent,
-    AddLabelComponent,
-    AppComponent,
-    ApplicationSortTableRowsComponent,
-    CommentPeriodBannerComponent,
-    CommentPeriodComponent,
-    CommentPeriodDetailsTabComponent,
-    CommentPeriodsComponent,
-    CommentPeriodsTableRowsComponent,
-    ConfirmComponent,
-    ContactsComponent,
-    DayCalculatorModalComponent,
-    DocumentApplicationSortComponent,
-    DocumentDetailComponent,
-    DocumentEditComponent,
-    DocumentTableRowsComponent,
-    FooterComponent,
-    GroupContactComponent,
-    GroupContactSelectComponent,
-    GroupTableRowsComponent,
-    GroupsTableRowsComponent,
-    HeaderComponent,
-    HomeComponent,
-    InputModalComponent,
-    LoginComponent,
-    MapComponent,
-    MetricsComponent,
-    MilestonesComponent,
-    NotAuthorizedComponent,
-    OrganizationsComponent,
-    OrganizationsTableRowsComponent,
-    PinsListComponent,
-    PinsTableRowsComponent,
-    ProjectArchivedDetailComponent,
-    ProjectComponent,
-    ProjectDetailComponent,
-    ProjectDocumentsComponent,
-    ProjectGroupsComponent,
-    ProjectNotificationTableRowsComponent,
-    ProjectNotificationsComponent,
-    ProjectUpdatesComponent,
-    ReviewCommentComponent,
-    ReviewCommentsTabComponent,
-    ReviewCommentsTabTableRowsComponent,
-    SearchComponent,
-    SearchDocumentTableRowsComponent,
-    SearchHelpComponent,
-    SidebarComponent,
-    ToggleButtonComponent,
-    ProjectDocumentsUploadComponent,
-    UserTableRowsComponent,
-    ProjectNotificationUploadComponent,
-    ProjectNotificationDetailComponent,
-    ProjectNotificationComponent,
-    ProjectNotificationDocumentsComponent,
-    PnDocumentTableRowsComponent,
-    AddEditProjectComponent,
-    ProjectListComponent,
-    ProjectListTableRowsComponent,
-    FormTab2002Component,
-    FormTab2018Component,
-    ModificationsListTableRowsComponent
-  ],
-  bootstrap: [AppComponent]
+    declarations: [
+        ActivityComponent,
+        ActivityDetailTableRowsComponent,
+        ActivityTableRowsComponent,
+        AddCommentComponent,
+        AddDocumentComponent,
+        AddDocumentTableRowsComponent,
+        AddEditActivityComponent,
+        AddEditCommentPeriodComponent,
+        AddEditContactComponent,
+        AddEditOrganizationComponent,
+        AddEditProjectNotificationComponent,
+        AddLabelComponent,
+        AppComponent,
+        ApplicationSortTableRowsComponent,
+        CommentPeriodBannerComponent,
+        CommentPeriodComponent,
+        CommentPeriodDetailsTabComponent,
+        CommentPeriodsComponent,
+        CommentPeriodsTableRowsComponent,
+        ConfirmComponent,
+        ContactsComponent,
+        DayCalculatorModalComponent,
+        DocumentApplicationSortComponent,
+        DocumentDetailComponent,
+        DocumentEditComponent,
+        DocumentTableRowsComponent,
+        FooterComponent,
+        GroupContactComponent,
+        GroupContactSelectComponent,
+        GroupTableRowsComponent,
+        GroupsTableRowsComponent,
+        HeaderComponent,
+        HomeComponent,
+        InputModalComponent,
+        LoginComponent,
+        MapComponent,
+        MetricsComponent,
+        MilestonesComponent,
+        NotAuthorizedComponent,
+        OrganizationsComponent,
+        OrganizationsTableRowsComponent,
+        PinsListComponent,
+        PinsTableRowsComponent,
+        ProjectArchivedDetailComponent,
+        ProjectComponent,
+        ProjectDetailComponent,
+        ProjectDocumentsComponent,
+        ProjectGroupsComponent,
+        ProjectNotificationTableRowsComponent,
+        ProjectNotificationsComponent,
+        ProjectUpdatesComponent,
+        ReviewCommentComponent,
+        ReviewCommentsTabComponent,
+        ReviewCommentsTabTableRowsComponent,
+        SearchComponent,
+        SearchDocumentTableRowsComponent,
+        SearchHelpComponent,
+        SidebarComponent,
+        ToggleButtonComponent,
+        ProjectDocumentsUploadComponent,
+        UserTableRowsComponent,
+        ProjectNotificationUploadComponent,
+        ProjectNotificationDetailComponent,
+        ProjectNotificationComponent,
+        ProjectNotificationDocumentsComponent,
+        PnDocumentTableRowsComponent,
+        AddEditProjectComponent,
+        ProjectListComponent,
+        ProjectListTableRowsComponent,
+        FormTab2002Component,
+        FormTab2018Component,
+        ModificationsListTableRowsComponent
+    ],
+    imports: [
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        BrowserModule,
+        CommonModule,
+        EditorModule,
+        FormsModule,
+        HttpClientModule,
+        MatMenuModule,
+        MatSlideToggleModule,
+        MatTabsModule,
+        NgbModule,
+        NgSelectModule,
+        NzPaginationModule,
+        ReactiveFormsModule,
+        SharedModule,
+    ],
+    providers: [
+        ApiService,
+        CanDeactivateGuard,
+        CommentPeriodService,
+        CommentService,
+        ConfigService,
+        CookieService,
+        CurrencyPipe,
+        DecisionService,
+        DocumentService,
+        ExcelService,
+        KeycloakService,
+        NotificationProjectService,
+        ProjectService,
+        RecentActivityService,
+        SearchService,
+        SideBarService,
+        StorageService,
+        UserService,
+        {
+            provide: APP_INITIALIZER,
+            useFactory: initConfig,
+            deps: [ConfigService, KeycloakService],
+            multi: true
+        },
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: TokenInterceptor,
+            multi: true
+        },
+    ],
+    bootstrap: [AppComponent]
 })
 
 export class AppModule {
