@@ -36,68 +36,65 @@ import { PublishedPipe } from './pipes/published.pipe';
 import { VarDirective } from './utils/ng-var.directive';
 
 @NgModule({
-  imports: [
-    RouterModule,
-    BrowserModule,
-    MatSlideToggleModule,
-    MatSnackBarModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatMenuModule,
-    NgbModule,
-    NzPaginationModule,
-    NzButtonModule,
-  ],
-  declarations: [
-    CommentStatsComponent,
-    ContactSelectComponent,
-    ContactSelectTableRowsComponent,
-    DropdownTemplateComponent,
-    FileUploadComponent,
-    LinkOrganizationComponent,
-    LinkOrganizationTableRowsComponent,
-    ListConverterPipe,
-    SafeHtmlPipe,
-    NewlinesPipe,
-    ObjectFilterPipe,
-    OrderByPipe,
-    OrgNamePipe,
-    PublishedPipe,
-    TableDirective,
-    TableTemplateComponent,
-    VarDirective,
-    ExtensionComponent
-  ],
-  entryComponents: [
-    ContactSelectTableRowsComponent
-  ],
-  exports: [
-    CommentStatsComponent,
-    ContactSelectComponent,
-    ContactSelectTableRowsComponent,
-    DropdownTemplateComponent,
-    FileUploadComponent,
-    LinkOrganizationComponent,
-    LinkOrganizationTableRowsComponent,
-    ListConverterPipe,
-    SafeHtmlPipe,
-    MatSlideToggleModule,
-    MatSnackBarModule,
-    NewlinesPipe,
-    OrderByPipe,
-    OrgNamePipe,
-    PublishedPipe,
-    TableTemplateComponent,
-    VarDirective,
-  ],
-  providers: [
-    TableTemplateUtils,
-    TableDocumentTemplateUtils,
-    NavigationStackUtils,
-    Utils,
-    { provide: NZ_I18N, useValue: en_US }
-  ]
+    imports: [
+        RouterModule,
+        BrowserModule,
+        MatSlideToggleModule,
+        MatSnackBarModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatMenuModule,
+        NgbModule,
+        NzPaginationModule,
+        NzButtonModule,
+    ],
+    declarations: [
+        CommentStatsComponent,
+        ContactSelectComponent,
+        ContactSelectTableRowsComponent,
+        DropdownTemplateComponent,
+        FileUploadComponent,
+        LinkOrganizationComponent,
+        LinkOrganizationTableRowsComponent,
+        ListConverterPipe,
+        SafeHtmlPipe,
+        NewlinesPipe,
+        ObjectFilterPipe,
+        OrderByPipe,
+        OrgNamePipe,
+        PublishedPipe,
+        TableDirective,
+        TableTemplateComponent,
+        VarDirective,
+        ExtensionComponent
+    ],
+    exports: [
+        CommentStatsComponent,
+        ContactSelectComponent,
+        ContactSelectTableRowsComponent,
+        DropdownTemplateComponent,
+        FileUploadComponent,
+        LinkOrganizationComponent,
+        LinkOrganizationTableRowsComponent,
+        ListConverterPipe,
+        SafeHtmlPipe,
+        MatSlideToggleModule,
+        MatSnackBarModule,
+        NewlinesPipe,
+        OrderByPipe,
+        OrgNamePipe,
+        PublishedPipe,
+        TableTemplateComponent,
+        VarDirective,
+    ],
+    providers: [
+        TableTemplateUtils,
+        TableDocumentTemplateUtils,
+        NavigationStackUtils,
+        Utils,
+        { provide: NZ_I18N, useValue: en_US }
+    ]
 })
 
 export class SharedModule { }
